Feature/nicol

+++ src/main/java/controlador/ControladorCategorias.java
@@ -11,8 +11,11 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import javax.swing.JOptionPane;
 import modelo.DAO.CategoriaDAO;
 import modelo.POJO.Categoria;
+import modelo.POJO.Usuario;
 
 /**
  *
@@ -22,6 +25,7 @@ public class ControladorCategorias extends HttpServlet {
 
     CategoriaDAO dao = new CategoriaDAO();
     Categoria cate = new Categoria();
+
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -39,7 +43,7 @@ public class ControladorCategorias extends HttpServlet {
             out.println("<!DOCTYPE html>");
             out.println("<html>");
             out.println("<head>");
-            out.println("<title>Servlet ControladorCategorias</title>");            
+            out.println("<title>Servlet ControladorCategorias</title>");
             out.println("</head>");
             out.println("<body>");
             out.println("<h1>Servlet ControladorCategorias at " + request.getContextPath() + "</h1>");
@@ -60,6 +64,41 @@ public class ControladorCategorias extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+
+        int id_categorias = Integer.parseInt(request.getParameter("idCategoria"));
+        String nom_cate = request.getParameter("nomCate");
+
+        HttpSession session = request.getSession();
+        session.setAttribute("categoria_actual", id_categorias);
+        session.setAttribute("nom_categoria", nom_cate);
+
+        if (session.getAttribute("idUsuario") == null) {
+            response.sendRedirect("tablaBolsillos.jsp");
+
+        } else {
+
+            int id_usuario = (int) session.getAttribute("idUsuario");
+            int id_categoria = (int) session.getAttribute("categoria_actual");
+
+            Categoria cat = new Categoria();
+            cat = dao.sumarPresupuestoGasto(id_usuario, id_categoria);
+
+            
+            Usuario usu = new Usuario();
+            usu = dao.presupuestoDisponible(id_usuario);
+
+            long presupuesto_disponible = usu.getPresupuesto_total() - usu.getPresupuesto_disponible();
+
+            usu.setPresupuesto_disponible(presupuesto_disponible);
+
+            request.setAttribute("Usuario", usu);
+            request.setAttribute("Categoria", cat);
+
+            request.getRequestDispatcher("tablaBolsillos.jsp").forward(request, response);
+
+        }
+
+        //response.sendRedirect("tablaBolsillos.jsp");
     }
 
     /**
@@ -73,7 +112,8 @@ public class ControladorCategorias extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-              String accion = request.getParameter("accion");
+
+        String accion = request.getParameter("accion");
         switch (accion) {
             case "Listar":
 
@@ -85,6 +125,7 @@ public class ControladorCategorias extends HttpServlet {
             case "Editar":
                 String ide = request.getParameter("id");
                 Categoria cat = dao.ListarId(ide);
+
                 request.setAttribute("Categoria", cat);
                 request.getRequestDispatcher("editCategoria.jsp").forward(request, response);
                 break;
@@ -92,57 +133,51 @@ public class ControladorCategorias extends HttpServlet {
             case "Actualizar":
                 String id1 = request.getParameter("txtid");
                 String nom1 = request.getParameter("txtnom");
-                String presupuesto1 = request.getParameter("txtpresupuesto");
-                String gasto1 = request.getParameter("txtgasto");
 
                 int idint = Integer.parseInt(id1);
-                long presupuestlolong = Long.parseLong(presupuesto1);
-                long gastolong = Long.parseLong(gasto1);
-                
+
                 cate.setId_categoria(idint);
                 cate.setNombre_categoria(nom1);
-                cate.setPresupuesto_categoria(presupuestlolong);
-                cate.setGasto_categoria(gastolong);
-                
+
                 dao.actualizar(cate);
                 request.getRequestDispatcher("ControladorCategorias?accion=Listar").forward(request, response);
                 break;
-            case "Delete":
+            case "Eliminar":
                 String id2 = request.getParameter("id");
-                dao.delete(id2);
+                int num_bolsillos = dao.num_bolsillos(id2);
+                if (num_bolsillos > 0) {
+
+                    JOptionPane.showMessageDialog(null, "No se puede borrar debido a bolsillos asociados");
+                    System.out.println("No se puede borrar debido a bolsillos asociados");
+
+                } else {
+                    dao.delete(id2);
+                }
+
                 request.getRequestDispatcher("ControladorCategorias?accion=Listar").forward(request, response);
                 break;
 
             case "Nuevo":
                 request.getRequestDispatcher("agregarCategorias.jsp").forward(request, response);
                 break;
-                
-                
-            case "submit":
-                
+
+            case "Enviar":
+
                 String id = request.getParameter("txtid");
                 String nombres = request.getParameter("txtnombres");
-                String presupuesto = request.getParameter("txtpresupuesto");
-                String gasto = request.getParameter("txtgasto");
-                
+
                 int id3 = Integer.parseInt(id);
-                long presupuesto3 = Long.parseLong(presupuesto);
-                long gasto3 = Long.parseLong(gasto);
-                
+
                 cate.setId_categoria(id3);
                 cate.setNombre_categoria(nombres);
-                cate.setPresupuesto_categoria(presupuesto3);
-                cate.setGasto_categoria(gasto3);
-                
+
                 dao.agregar(cate);
                 request.getRequestDispatcher("ControladorCategorias?accion=Listar").forward(request, response);
-                
-                    
-                    break;
+
+                break;
             default:
                 throw new AssertionError();
-        }           
-
+        }
 
     }
